:recycle: refactor 글 등록 페이지 제외 1차 리팩토링
